terzo commit

+++ DynamicECommerce/DECommerce.Models/Products.cs
@@ -12,7 +12,7 @@ namespace DECommerce.Models
         [Key]
         public int ProductID { get; set; }
         public int ProductCategoriesID { get; set; }
-        public int? UnitPrice { get; set; }
+        public decimal? UnitPrice { get; set; }
         public string? Image { get; set; }
         public string? Field1 { get; set; }
         public string? Field2 { get; set; }
